Extract HTML helpers and guard parsing with Trusted Types
